+++ platform/native/drivers/dma_hal.h
@@ -5,7 +5,7 @@
 #include <stddef.h>
 
 void dma_hal_init(void *hal_handle, void *config_ptr);
-void dma_hal_start(void *hal_handle, uint32_t src, uint32_t dst, uint32_t length);
+void dma_hal_start(void *hal_handle, uintptr_t src, uintptr_t dst, size_t length);
 void dma_hal_stop(void *hal_handle);
 
 #endif /* DMA_HAL_NATIVE_H */
